Add performance measurement

+++ include/primitives/radix-sort.hh
@@ -1,4 +1,7 @@
 #pragma once
+
+#include <cstdint>
+
 struct GL;
-void radix_sort(GL const & gl, unsigned int key, unsigned int size, unsigned int index = 0,
+void radix_sort(GL const & gl, unsigned int key, ptrdiff_t size, unsigned int index = 0,
   bool descending = false, bool is_signed = false, bool is_float = false);
